requirements/local.txt
django-mysql
mysqlclient
